requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
pypng==0.20220715.0
python-decouple==3.8
qrcode==7.4.2
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
wheel==0.44.0
